tests(testing): add missing async client tests
