+++ tests/unit/test_contrib/test_sqlalchemy/test_repository/test_repo_bigint.py
@@ -23,6 +23,7 @@ from litestar.contrib.repository.filters import (
     SearchFilter,
 )
 from litestar.contrib.sqlalchemy import base
+from tests.helpers import maybe_async
 from tests.unit.test_contrib.test_sqlalchemy.models_bigint import (
     AuthorAsyncRepository,
     AuthorSyncRepository,
@@ -44,7 +45,7 @@ from tests.unit.test_contrib.test_sqlalchemy.models_bigint import (
     TagSyncRepository,
 )
 
-from .helpers import maybe_async, update_raw_records
+from .helpers import update_raw_records
 
 
 @pytest.fixture()
